+++ iivs/dhm/data/intensity/__init__.py
@@ -5,8 +5,12 @@ __all__ = (
     "IntensityBinHeader",
     "IntensityBinList",
     "IntensitySequence",
+    "IntensityTxtFolder",
+    "IntensityTxtList",
     "load_intensity_bin",
+    "load_intensity_txt",
     "read_intensity_bin_header",
+    "read_intensity_txt_header",
     "save_intensity_bin",
 )
 
@@ -18,4 +22,10 @@ from iivs.dhm.data.intensity.bin import (
     load_intensity_bin,
     read_intensity_bin_header,
     save_intensity_bin,
+)
+from iivs.dhm.data.intensity.txt import (
+    IntensityTxtFolder,
+    IntensityTxtList,
+    load_intensity_txt,
+    read_intensity_txt_header,
 )