+++ iivs/dhm/data/phase/__init__.py
@@ -5,10 +5,14 @@ __all__ = (
     "PhaseBinHeader",
     "PhaseBinList",
     "PhaseSequence",
+    "PhaseTxtFolder",
+    "PhaseTxtList",
     "PhaseUnit",
     "convert_phase_unit",
     "load_phase_bin",
+    "load_phase_txt",
     "read_phase_bin_header",
+    "read_phase_txt_header",
     "save_phase_bin",
 )
 
@@ -22,3 +26,9 @@ from iivs.dhm.data.phase.bin import (
     save_phase_bin,
 )
 from iivs.dhm.data.phase.core import PhaseUnit, convert_phase_unit
+from iivs.dhm.data.phase.txt import (
+    PhaseTxtFolder,
+    PhaseTxtList,
+    load_phase_txt,
+    read_phase_txt_header,
+)